sequelize.import
Hi, I've been following the steps in this guide to add a currency system to my bot and i've noticed a problem. With the latest uptdate of sequelize the sequelize.import does'nt work. I've managed to solve this with a require(path.join()) type of code instead, it's pretty easy to find on stackoverflow.